refactor: combined a bunch of code that didnt need to be separate. hoping this makes later changes easier
