--- Assets/Code/MissionControl.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bda3a3c7ec8cbcd4b84bd033f96f47f4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
