--- Assets/Code/UiUpdater.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: acb1a3c170a764f448b566faecc2d196
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
