--- Assets/Code/YearAdvancer.cs
@@ -1,14 +1,0 @@
-﻿using UnityEngine;
-using UnityEngine.UI;
-
-namespace TakeTheSky
-{
-    public class YearAdvancer : MonoBehaviour
-    {
-        public void EndYear()
-        {
-            CurrentState.CurrentYear++;
-            CurrentState.CurrentEp += CurrentState.EpGainPerYear;
-        }
-    }
-}
